Meta: Some compatibility fixes: Ninja has problem with $ signs, add include of the gendatadir to improve compatibility to current serenity repo
For example the LibAudio includes the generated file <AudioServer/AudioClientEndpoint.h>. This file cannot be found with the meta
build system, because the file is generated for the LibAudio itself, thus the include path must be <AudioClientEndpoint.h> or "AudioClientEndpoint.h".
As the file is existent in the gendatadir/AudioServer/AudioClientEndpoint.h it is sufficient to have the gendatadir in the include-dirs, so the
file can be found without changing anything within the serenityos repository.