Meta: Support compile definitions from output of executed commands
This helps filling the GIT_* definitions used in the serenity's About
application. this is done via an object of "DEFINITION": "COMMAND" within
the execution_result_definitions item in host_tools/tool below a package.
Also, the clearing of the cmake build flags has been separated from
setting custom flags on a target. This can be achieved by setting the
"with the reset_toolchain_flags" item to true.