Meta: Support different machine types in one package description
This allows to specify common attributes for a package in the main section
and machine specific attributes within the machine section.

An example to demonstrate this is LibCore, where dependency and sources change
for specific machine, the include directory however applies to both:

            "machine": {
                "host": {
                    "source": [
                        "${root}/Libraries/LibCore/IODevice.cpp",
                        "${root}/Libraries/LibCore/File.cpp",
                        "${root}/Libraries/LibCore/Object.cpp",
                        ...
                    ]
                },
                "target": {
                    "source": [
                        "${root}/Libraries/LibCore/*.cpp",
                        "${root}/Libraries/LibCore/*.c"
                    ],
                    "dependency": [
                        "LibC"
                    ]
                }
            },
            "include": [
                "${root}/Libraries/LibCore"
            ],

This patchset makes the suffixes "_build" or "_host" for the packages that have
been used so far obsolete.